requirements.txt
asgiref==3.8.1
boto3==1.35.22
botocore==1.35.22
Django==5.0.4
django-appconf==1.0.6
django-cleanup==8.1.0
django-countries==7.6.1
django-js-asset==2.2.0
django-storages==1.14.4
jmespath==1.0.1
pillow==10.3.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
rcssmin==1.1.1
rjsmin==1.2.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.4.4
toposort==1.10
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.3
